Phase 13: use the reference's single-precision constants (#349)
* Use the reference's single-precision constants

The reference writes its density normalizers as default-kind literals
widened with dble, so the binary uses their float32 roundings. All three
backends now read those values, and epsdble's, from one shared module,
pamica/reference_constants.py; MLX holds the rho == 2 normalizer in its
lgamma table. The literal-formula tests and the benchmark's copy now use
the reference's values. Tested: pdf-family and backend test files pass.

* Compare pre-change fits at the old constants

The byte-identity tests against code from before issue #344 (component
rows, doscaling off, the MLX pre-Phase-3 tip) now give the live backend
its old density constants through pre_change.use_pre_344_constants, so
they still isolate the change each was written for: their two-model and
non-GG fits reach the changed normalizers. Tested: the three files pass.

* Pin the constants against the reference

test_reference_constants.py checks each shared constant against the
float32 rounding of its literal on the cited reference line (exact
rational arithmetic), pins the sweep of amica15.f90 and its header,
checks that no backend keeps a copy, checks every backend's rho == 2
log-density, and (opt-in) seeds the native binary for pdftype 2, 4 and
1. Tested: 23 passed with AMICA_RUN_FORTRAN=1.

* Run the merged oracle at rho == 2

The merged-state oracle no longer holds maxrho at 1.99: its warm seed
now has mixtures clamped at 2, it asserts that, and the code before
issue #344 is replayed from the same state as a control (off by 2.8e-9
in log-likelihood after 1 iteration, against 2.2e-15 now). Tolerances
unchanged. Tested with AMICA_RUN_FORTRAN=1.

* Document the single-precision constants

Changelog entry with the oracle numbers; the differences guide gains a
section on the constants, and row 16 now records the one kind of
single-precision literal pamica keeps at its decimal value (defaults of
input.param keys); the validation guide notes the new family oracle.
Tested: mkdocs build --strict, typos.

* Note the reference's unset lrate0 in row 16

Without an lrate key the reference never sets its ceiling lrate0, so
its lrate drops to 0 after the first iteration (checked against the
pinned binary). Row 16 now says so and uses comp_thresh as the example.
Tested: mkdocs build --strict, typos.

* Skip MLX before patching its constants

On a host without MLX the byte-identity tests imported the MLX module
to patch it before the check that skips them. Resolve the classes
first. Tested: the two files pass with MLX and skip its cases without.

* Remove the unused compute_log_pdf

numpy_impl.pdf.compute_log_pdf had no callers in the package, tests or
docs. Tested: the pdf and viz tests pass.

* Plot the fit's density in compute_pdf

compute_pdf, which viz.plot_pdf_fits draws, now takes its normalizers
from pamica.reference_constants, including the reference's
single-precision sqrt(pi) at rho == 2, so it draws the density the fit
uses. Its docstring says it is a plotting helper with the legacy
pdftype numbering. Tested: the pdf, viz and constants tests pass.

* Scan every backend module for copies

The no-copy test now scans every .py file of torch_impl, numpy_impl and
mlx_impl with comments and strings blanked, and flags the linear forms
(sqrt(pi), sqrt(2*pi), ...) as well as the log forms and the literals.
A new test checks that it flags every definition this change removed.
Tested: passes; fails on the old pdf.py sqrt(pi) line (checked by hand).

* Assert the family held in the oracle

The family oracle now asserts that every source kept its family and no
kurtosis switch ran (n_kurt_done == 0), for pdftype=1 with num_kurt=0
too, in the live and the pre-change runs. Tested with
AMICA_RUN_FORTRAN=1.

* Check the pre-#344 constant substitution

A control test: every constant use_pre_344_constants patches differs
from the live value, the patch takes effect, it covers every changed
constant each backend module reads, and on MLX it restores the
pre-change rho == 2 table entry. Tested: passes on all three backends.

* Note the plotting helper in the changelog

compute_pdf now draws the fit's density and compute_log_pdf is gone.
Tested: mkdocs build --strict, typos.

+++ benchmarks/reproduce_table1.py
@@ -595,10 +595,12 @@ def run_multimodel_ensemble(
 # (matches test_ng_backend.py::test_sufficient_stats_match_numpy_reference).
 # ---------------------------------------------------------------------------
 
+# The reference's log-normalizers: each default-kind literal is rounded to
+# single precision before ``dble`` widens it (issue #344); 4.0 is exact.
 _LOG4 = math.log(4.0)
-_LSQ2PI = math.log(2.506628274)
-_LNSUB = math.log(4.132731354)
-_LNSUP = math.log(1.858073988)
+_LSQ2PI = math.log(float(np.float32(2.506628274)))
+_LNSUB = math.log(float(np.float32(4.132731354)))
+_LNSUP = math.log(float(np.float32(1.858073988)))
 
 
 def _fortran_z0(y: np.ndarray, code: int) -> np.ndarray:
